+++ supersuit/__init__.py
@@ -7,11 +7,10 @@ __version__ = "2.6.6"
 
 
 from .generic_wrappers import * # NOQA
+from .lambda_wrappers import action_lambda_v1, observation_lambda_v0, reward_lambda_v0
 from .multiagent_wrappers import agent_indicator_v0, black_death_v1, black_death_v0, \
     pad_action_space_v0, pad_observations_v0
-
-from supersuit.generic_wrappers import frame_skip_v0, action_lambda_v1, observation_lambda_v0, \
-    reward_lambda_v0, color_reduction_v0, resize_v0, dtype_v0, \
+from supersuit.generic_wrappers import frame_skip_v0, color_reduction_v0, resize_v0, dtype_v0, \
     flatten_v0, reshape_v0, normalize_obs_v0, clip_actions_v0, clip_reward_v0, \
     delay_observations_v0, frame_stack_v1, frame_stack_v0, max_observation_v0, \
     sticky_actions_v0
