+++ supersuit/generic_wrappers/__init__.py
@@ -1,9 +1,8 @@
 from .frame_skip import frame_skip_v0
-from .lambda_wrappers import action_lambda_v1, observation_lambda_v0, reward_lambda_v0
 from .basic_wrappers import color_reduction_v0, resize_v0, dtype_v0, \
     flatten_v0, reshape_v0, normalize_obs_v0, clip_actions_v0, clip_reward_v0
 
-from .modifier_wrappers.delay_observations import delay_observations_v0
-from .modifier_wrappers.frame_stack import frame_stack_v1, frame_stack_v0
-from .modifier_wrappers.max_observation import max_observation_v0
-from .modifier_wrappers.sticky_actions import sticky_actions_v0
+from .delay_observations import delay_observations_v0
+from .frame_stack import frame_stack_v1, frame_stack_v0
+from .max_observation import max_observation_v0
+from .sticky_actions import sticky_actions_v0
