+++ supersuit/generic_wrappers/max_observation.py
